Group packages 'key', 'keyboard', 'layout' and 'smartbar' into 'text'
This change is necessary to make room for the implementation of emojis,
gif picker, etc.

This keyboard is now divided into the following parts:
 'text' with 'TextInputManager' as controller
 'media' with 'MediaInputManager' as controller

Dividing the keyboard into these two parts helps with keeping the
relative simple stuff like entering text from the more complicated
inputting media into application editors. The smartbar will also be a
'text'-only bar, as it serves no purpose for doing anything helpful
while browsing emojis. At its place in the UI will be a search bar to
search for gifs, emojis (by name), etc.

Note, that the manager classes and the 'media' package are not yet
implemented, but in final planning phase.