Up (#4)
* Add AI_POLICY.md (#3133)

* Add AI_POLICY.md

* Add AI policy to issue form templates

* Add pull request template

---------

Co-authored-by: Patrick Goldinger <patrick@patrickgold.dev>

* Implement 'Hide Keyboard' action (#3132)

* Feat: Implement 'Hide Keyboard' action

Adds a new utility function that dismisses the active soft keyboard when triggered from the Quick Action menu. This improves user flow when managing quick settings.

* Feat: Implement 'Hide Keyboard' action

Adds a new utility function that dismisses the active soft keyboard when triggered from the Quick Action menu. This improves user flow when managing quick settings.

* Fix uppercase letters missing in email regex (#3110)

* Update Gradle to 9.2.0 and AGP to 8.12.0

* Update Kotlin to 2.2.20

* Update androidx dependencies

* Update aboutlibraries to 13.1.0

* Migrate to type-safe project accessors

* Update Jetpref

* Properly recalculate SurfaceView size (#3040)

This commit fixes the issues of the SurfaceView size change.
Previusly the SurfaceView would not change its size when the
keyboard was rezised which caused transparent areas in the
background.

* Add default Material You themes (#3135)

* Add default Material You themes

* Move Material You themes to their own extension

---------

Co-authored-by: Lars Mühlbauer <lm41@lm41.xyz>
Co-authored-by: Patrick Goldinger <patrick@patrickgold.dev>
Co-authored-by: Quy Cao <quycao@live.com>
Co-authored-by: klaurence <6565795+klaurence@users.noreply.github.com>